feat: allow single image only (#6)
* chore: remove husky temporarily

Remove husky as it was failing to commit multiple times, will need to look into why.

* feat: start adding ability to update single image

Add some debug lines to pull the current comments on a PR, this is to start making it so we can
update a single comment with new information

* feat: begin to try to update the comment if we see an old comment

If we see an old comment defined by an invisible tag, then try to update it. For now this seems the
best way to ensure that it is unique.

* feat: add new flag to control if it should be an update

Add new flag to denote if the comment should be updated on each execution of the github action, or
if a new comment should be created every time its run.